fixed clantag auth and added new arxan checks
The string tables are now properly checked, ensuring access is granted to the appropriate users.
Added functionality to allow the use of all clantags, except for elite tags, even when no Discord UID is present.
Introduced additional window checks for Arxan and created a separate file to maintain a cleaner structure.